backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.32.4
python-dotenv==1.0.1
anthropic==0.68.0
pydantic==2.10.0
beautifulsoup4==4.12.3
